ENH: Change size, index and offset definitions: use size_t and ptrdiff_t
Instead of unsigned long and long. Adapt surrounding code to use
specific types where possible (i.e., PointIdentifier, CellIdentifier)
but to use size_t and ptrdiff_t where intent is less than crystal
clear.

This commit focuses on changes in the Code/Common directory, and
associated changes in other directories required to build and test
the changes. The remaining directories still need a sweep through
to eliminate inappropriate usage of "long" and "unsigned long".

This commit enables the use of image sizes, offsets and indexes
larger than 2GB on Win64.

For 32-bit Windows, and for 32-bit and 64-bit Linux, this change
should largely be a no-op. The types involved correlate to the
new types being used.

I chose using size_t and ptrdiff_t so that these builds would remain
working as is, including client code, maximizing backwards compatibility.

Code modifications in Testing and Examples are only needed for Win64 builds.
The Testing and Examples would build without code modifications on non-Win64
builds.

+++ Code/Algorithms/itkSimplexMeshVolumeCalculator.h
@@ -110,7 +110,7 @@ public:
     /**
      * \brief visits all polygon cells and compute the cell centers
      */
-    void Visit(unsigned long cellId, SimplexPolygonType *poly)
+    void Visit(size_t cellId, SimplexPolygonType *poly)
     {
       typedef typename SimplexPolygonType::PointIdIterator PointIdIterator;
       PointIdIterator it =  poly->PointIdsBegin();
