+++ Code/Common/itkMeshRegion.h
@@ -69,31 +69,31 @@ public:
   { return Superclass::ITK_UNSTRUCTURED_REGION; }
 
   /** Get the number of regions. */
-  unsigned long GetNumberOfRegions() const
+  size_t GetNumberOfRegions() const
   { return m_NumberOfRegions; }
 
   /** Set the number of regions. */
-  void SetNumberOfRegions(unsigned long num)
+  void SetNumberOfRegions(size_t num)
   {
-    if ( ( num >= 1 ) && ( num <= NumericTraits< unsigned long >::max() ) )
+    if ( ( num >= 1 ) && ( num <= NumericTraits< size_t >::max() ) )
               { m_NumberOfRegions = num; } }
 
   /** Get the current region. */
-  unsigned long GetRegion() const
+  size_t GetRegion() const
   { return m_Region; }
 
   /** Set the number of regions. */
-  void SetRegion(unsigned long region)
+  void SetRegion(size_t region)
   {
-    if ( ( region >= 1 ) && ( region <= NumericTraits< unsigned long >::max() ) )
+    if ( ( region >= 1 ) && ( region <= NumericTraits< size_t >::max() ) )
               { m_Region = region; } }
 
 private:
   // The maximum number of regions possible.
-  unsigned long int m_NumberOfRegions;
+  size_t m_NumberOfRegions;
 
   // The specified region.
-  unsigned long int m_Region;
+  size_t m_Region;
 };
 } // end namespace itk
